# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/questions": {
      "post": {
        "operationId": "QuestionsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateQuestionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionDto"
                }
              }
            }
          }
        },
        "summary": "Создать вопрос",
        "tags": [
          "questions"
        ]
      },
      "get": {
        "operationId": "QuestionsController_findAll",
        "parameters": [
          {
            "name": "difficulty",
            "required": false,
            "in": "query",
            "schema": {
              "example": "middle",
              "type": "string",
              "enum": [
                "junior",
                "middle",
                "senior",
                "lead"
              ]
            }
          },
          {
            "name": "topic",
            "required": false,
            "in": "query",
            "description": "Имя темы (name или slug)",
            "schema": {
              "example": "js",
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "default": 0,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/QuestionDto"
                  }
                }
              }
            }
          }
        },
        "summary": "Список вопросов (CRUD-выборка)",
        "tags": [
          "questions"
        ]
      }
    },
    "/api/questions/export": {
      "get": {
        "operationId": "QuestionsController_exportCsv",
        "parameters": [
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Поисковый запрос",
            "schema": {
              "example": "scope",
              "type": "string"
            }
          },
          {
            "name": "difficulty",
            "required": false,
            "in": "query",
            "description": "Фильтр по сложности (CSV или repeatable query)",
            "schema": {
              "example": [
                "junior",
                "middle"
              ],
              "type": "array",
              "items": {
                "type": "string",
                "enum": [
                  "junior",
                  "middle",
                  "senior",
                  "lead"
                ]
              }
            }
          },
          {
            "name": "topicIds",
            "required": false,
            "in": "query",
            "description": "Фильтр по topicId (CSV или repeatable query)",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "companyQuery",
            "required": false,
            "in": "query",
            "description": "Строковый поиск по названию компании",
            "schema": {
              "example": "google",
              "type": "string"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "schema": {
              "default": "relevance",
              "type": "string",
              "enum": [
                "relevance",
                "newest",
                "popular"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "text/csv": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        },
        "summary": "Экспорт вопросов в CSV",
        "tags": [
          "questions"
        ]
      }
    },
    "/api/questions/{id}": {
      "get": {
        "operationId": "QuestionsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionDto"
                }
              }
            }
          }
        },
        "summary": "Получить вопрос по id",
        "tags": [
          "questions"
        ]
      },
      "patch": {
        "operationId": "QuestionsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateQuestionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionDto"
                }
              }
            }
          }
        },
        "summary": "Обновить вопрос",
        "tags": [
          "questions"
        ]
      },
      "delete": {
        "operationId": "QuestionsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "summary": "Удалить вопрос",
        "tags": [
          "questions"
        ]
      }
    },
    "/api/questions/import/preview": {
      "post": {
        "operationId": "QuestionsController_previewImport",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "file"
                ],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionCsvImportReportDto"
                }
              }
            }
          }
        },
        "summary": "Preview CSV import for questions",
        "tags": [
          "questions"
        ]
      }
    },
    "/api/questions/import/commit": {
      "post": {
        "operationId": "QuestionsController_commitImport",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "file"
                ],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionCsvImportReportDto"
                }
              }
            }
          }
        },
        "summary": "Commit CSV import for questions",
        "tags": [
          "questions"
        ]
      }
    },
    "/api/questions/{id}/interview-encounter": {
      "put": {
        "operationId": "QuestionsController_markInterviewEncounter",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionInterviewEncounterDto"
                }
              }
            }
          }
        },
        "summary": "Отметить вопрос как встреченный на собеседовании",
        "tags": [
          "questions"
        ]
      },
      "delete": {
        "operationId": "QuestionsController_unmarkInterviewEncounter",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionInterviewEncounterDto"
                }
              }
            }
          }
        },
        "summary": "Снять отметку вопроса как встреченного на собеседовании",
        "tags": [
          "questions"
        ]
      }
    },
...
    "/api/question-change-requests": {
      "post": {
        "operationId": "QuestionChangeRequestsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateQuestionChangeRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionChangeRequestDetailDto"
                }
              }
            }
          }
        },
        "summary": "Создать заявку на изменение вопроса",
        "tags": [
          "question-change-requests"
        ]
      }
    },
...
    "/api/question-change-requests/{id}": {
      "get": {
        "operationId": "QuestionChangeRequestsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionChangeRequestDetailDto"
                }
              }
            }
          }
        },
        "summary": "Детали заявки на изменение вопроса",
        "tags": [
          "question-change-requests"
        ]
      }
    },
    "/api/question-change-requests/{id}/approve": {
      "post": {
        "operationId": "QuestionChangeRequestsController_approve",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionChangeRequestDetailDto"
                }
              }
            }
          }
        },
        "summary": "Одобрить заявку",
        "tags": [
          "question-change-requests"
        ]
      }
    },
    "/api/question-change-requests/{id}/reject": {
      "post": {
        "operationId": "QuestionChangeRequestsController_reject",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectQuestionChangeRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionChangeRequestDetailDto"
                }
              }
            }
          }
        },
        "summary": "Отклонить заявку",
        "tags": [
          "question-change-requests"
        ]
      }
    },
...
  },
  "info": {
    "title": "Nord API",
    "description": "Backend API for interview questions bank",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "QuestionStructuredContentDto": {
        "type": "object",
        "properties": {
          "text": {
            "type": "string",
            "example": "Что выведет этот код?",
            "maxLength": 5000
          },
          "code": {
            "type": "string",
            "example": "const answer: number = 42;",
            "maxLength": 5000
          },
          "codeLanguage": {
            "type": "string",
            "enum": [
              "javascript",
              "jsx",
              "typescript",
              "tsx"
            ],
            "example": "typescript"
          }
        },
        "required": [
          "text"
        ]
      },
      "QuestionTopicDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "cm7y5z8qv0003x4f2w7sn1abc"
          },
          "name": {
            "type": "string",
            "example": "JS"
          },
          "slug": {
            "type": "string",
            "example": "js"
          }
        },
        "required": [
          "id",
          "name",
          "slug"
        ]
      },
...
      "CreateQuestionDto": {
        "type": "object",
        "properties": {
          "textContent": {
            "description": "Структурированное содержимое вопроса",
            "allOf": [
              {
                "$ref": "#/components/schemas/QuestionStructuredContentDto"
              }
            ]
          },
          "answerContent": {
            "description": "Структурированное содержимое ответа",
            "allOf": [
              {
                "$ref": "#/components/schemas/QuestionStructuredContentDto"
              }
            ]
          },
          "difficulty": {
            "type": "string",
            "description": "Сложность вопроса",
            "enum": [
              "junior",
              "middle",
              "senior",
              "lead"
            ],
            "example": "junior"
          },
          "topicIds": {
            "description": "Список topicId",
            "minItems": 1,
            "example": [
              "cm7y5z8qv0003x4f2w7sn1abc"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "companyId": {
            "type": "string",
            "description": "Опциональный companyId, если вопрос встречался на собеседовании в компании",
            "nullable": true,
            "example": "cm7y5z8qv0003x4f2w7sn1company"
          }
        },
        "required": [
          "textContent",
          "answerContent",
          "difficulty",
          "topicIds"
        ]
      },
      "QuestionCompanyDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "cm7y5z8qv0003x4f2w7sn1abc"
          },
          "name": {
            "type": "string",
            "example": "Google"
          }
        },
        "required": [
          "id",
          "name"
        ]
      },
      "QuestionPendingChangeRequestDto": {
        "type": "object",
        "properties": {
          "hasPendingChangeRequest": {
            "type": "boolean",
            "example": true
          },
          "hasMyPendingChangeRequest": {
            "type": "boolean",
            "example": false
          }
        },
        "required": [
          "hasPendingChangeRequest",
          "hasMyPendingChangeRequest"
        ]
      },
      "QuestionInterviewEncounterDto": {
        "type": "object",
        "properties": {
          "count": {
            "type": "number",
            "example": 12
          },
          "checkedByCurrentUser": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "count",
          "checkedByCurrentUser"
        ]
      },
      "QuestionDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "cm7y5z8qv0001x4f2w7sn1aaa"
          },
          "text": {
            "type": "string",
            "example": "Какие основные типы данных есть в JavaScript?"
          },
          "textContent": {
            "$ref": "#/components/schemas/QuestionStructuredContentDto"
          },
          "answer": {
            "type": "string",
            "example": "string, number, boolean, null, undefined, symbol, bigint, object."
          },
          "answerContent": {
            "$ref": "#/components/schemas/QuestionStructuredContentDto"
          },
          "difficulty": {
            "type": "string",
            "enum": [
              "junior",
              "middle",
              "senior",
              "lead"
            ],
            "example": "junior"
          },
          "company": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/QuestionCompanyDto"
              }
            ]
          },
          "topics": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QuestionTopicDto"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "pendingChangeRequest": {
            "$ref": "#/components/schemas/QuestionPendingChangeRequestDto"
          },
          "interviewEncounter": {
            "$ref": "#/components/schemas/QuestionInterviewEncounterDto"
          }
        },
        "required": [
          "id",
          "text",
          "textContent",
          "answer",
          "answerContent",
          "difficulty",
          "company",
          "topics",
          "createdAt",
          "updatedAt",
          "pendingChangeRequest",
          "interviewEncounter"
        ]
      },
      "UpdateQuestionDto": {
        "type": "object",
        "properties": {
          "textContent": {
            "description": "Структурированное содержимое вопроса",
            "allOf": [
              {
                "$ref": "#/components/schemas/QuestionStructuredContentDto"
              }
            ]
          },
          "answerContent": {
            "description": "Структурированное содержимое ответа",
            "allOf": [
              {
                "$ref": "#/components/schemas/QuestionStructuredContentDto"
              }
            ]
          },
          "difficulty": {
            "type": "string",
            "description": "Сложность вопроса",
            "enum": [
              "junior",
              "middle",
              "senior",
              "lead"
            ],
            "example": "middle"
          },
          "topicIds": {
            "description": "Список topicId",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "companyId": {
            "type": "string",
            "description": "Новый companyId или null для отвязки компании",
            "nullable": true,
            "example": "cm7y5z8qv0003x4f2w7sn1company"
          }
        }
      },
...
      "QuestionCsvImportReportDto": {
        "type": "object",
        "properties": {
          "applied": {
            "type": "boolean",
            "example": false
          },
          "fileName": {
            "type": "string",
            "example": "_Вопросы - Основной.csv"
          },
          "delimiter": {
            "type": "string",
            "enum": [
              ";",
              ","
            ],
            "example": ";"
          },
          "warnings": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "topicsToCreate": {
            "example": [
              "Angular",
              "Browser APIs"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "companiesToCreate": {
            "example": [
              "Google",
              "Yandex"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "totals": {
            "$ref": "#/components/schemas/QuestionCsvImportTotalsDto"
          },
          "rows": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QuestionCsvImportRowDto"
            }
          }
        },
        "required": [
          "applied",
          "fileName",
          "delimiter",
          "warnings",
          "topicsToCreate",
          "companiesToCreate",
          "totals",
          "rows"
        ]
      },
...
      "CreateQuestionChangeRequestDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "CREATE",
              "UPDATE",
              "DELETE"
            ],
            "example": "UPDATE"
          },
          "targetQuestionId": {
            "type": "string",
            "description": "ID опубликованного вопроса для UPDATE или DELETE",
            "example": "cm7y5z8qv0001x4f2w7sn1aaa"
          },
          "payload": {
            "$ref": "#/components/schemas/QuestionChangeRequestPayloadDto"
          }
        },
        "required": [
          "type"
        ]
      },
...
      "QuestionChangeRequestDetailDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "cm8q4x7r10001s2t3u4v5w6x7"
          },
          "type": {
            "type": "string",
            "enum": [
              "CREATE",
              "UPDATE",
              "DELETE"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "APPROVED",
              "REJECTED"
            ]
          },
          "targetQuestionId": {
            "type": "object",
            "nullable": true,
            "example": "cm7y5z8qv0001x4f2w7sn1aaa"
          },
          "subject": {
            "type": "string",
            "example": "Что такое MVCC?"
          },
          "author": {
            "$ref": "#/components/schemas/QuestionChangeRequestActorDto"
          },
          "reviewer": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/QuestionChangeRequestActorDto"
              }
            ]
          },
          "reviewComment": {
            "type": "object",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "reviewedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "before": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/QuestionSnapshotDto"
              }
            ]
          },
          "after": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/QuestionSnapshotDto"
              }
            ]
          },
          "fieldDiffs": {
            "$ref": "#/components/schemas/QuestionChangeRequestFieldDiffsDto"
          }
        },
        "required": [
          "id",
          "type",
          "status",
          "targetQuestionId",
          "subject",
          "author",
          "reviewer",
          "reviewComment",
          "createdAt",
          "updatedAt",
          "reviewedAt",
          "before",
          "after",
          "fieldDiffs"
        ]
      },
...
      "RejectQuestionChangeRequestDto": {
        "type": "object",
        "properties": {
          "reviewComment": {
            "type": "string",
            "description": "Причина отклонения",
            "maxLength": 2000,
            "example": "Слишком общая формулировка и дублирующийся тег."
          }
        },
        "required": [
          "reviewComment"
        ]
      },
...
    }
  }
}